layout_TableLayoutのAndroid Studio対応 #7
