+++ Chapter01/layout_TableLayout/.gitignore
@@ -1,0 +1,6 @@
+.gradle
+/local.properties
+/.idea/workspace.xml
+/.idea/libraries
+.DS_Store
+/build
