+++ Chapter01/layout_TableLayout/.idea/modules.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/layout_TableLayout.iml" filepath="$PROJECT_DIR$/layout_TableLayout.iml" />
+    </modules>
+  </component>
+</project>
+
