+++ Chapter01/layout_TableLayout/app/src/androidTest/java/net/buildbox/pokeri/layout_tablelayout/ApplicationTest.java
@@ -1,0 +1,13 @@
+package net.buildbox.pokeri.layout_tablelayout;
+
+import android.app.Application;
+import android.test.ApplicationTestCase;
+
+/**
+ * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
+ */
+public class ApplicationTest extends ApplicationTestCase<Application> {
+    public ApplicationTest() {
+        super(Application.class);
+    }
+}
