+++ Chapter01/layout_TableLayout/settings.gradle
@@ -1,0 +1,1 @@
+include ':app'
