--- Chapter01/layout_TableLayout/src/net/buildbox/pokeri/layout_tablelayout/MainActivity.java
@@ -1,21 +1,0 @@
-package net.buildbox.pokeri.layout_tablelayout;
-
-import android.os.Bundle;
-import android.app.Activity;
-import android.view.Menu;
-
-public class MainActivity extends Activity {
-
-	@Override
-	protected void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-		setContentView(R.layout.activity_main);
-	}
-
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		getMenuInflater().inflate(R.menu.main, menu);
-		return true;
-	}
-
-}
